Add indexes, limits and resolve nits
