+++ analytics-writer/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@filbeam/analytics-worker",
+  "name": "@filbeam/analytics-writer",
   "version": "1.0.0",
   "private": true,
   "type": "module",
