+++ analytics-writer/worker-configuration.d.ts
@@ -1,13 +1,12 @@
 /* eslint-disable */
-// Generated by Wrangler by running `wrangler types` (hash: d9d95d8524cf70ca24d7d54944858a17)
-// Runtime types generated with workerd@1.20251004.0 2024-01-01 nodejs_compat
+// Generated by Wrangler by running `wrangler types` (hash: 7582f67fbcdf1b93f808f20b163c8a39)
+// Runtime types generated with workerd@1.20251008.0 2024-01-01 nodejs_compat
 declare namespace Cloudflare {
 	interface GlobalProps {
-		mainModule: typeof import("./bin/analytics-worker");
+		mainModule: typeof import("./bin/analytics-writer");
 	}
 	interface Env {
-		ENVIRONMENT: "dev" | "calibration" | "mainnet";
-		"analytics-engine": AnalyticsEngineDataset;
+		analytics_engine: AnalyticsEngineDataset;
 	}
 }
 interface Env extends Cloudflare.Env {}
