+++ vitest.workspace.js
@@ -1,9 +1,10 @@
 import { defineWorkspace } from 'vitest/config'
 
 export default defineWorkspace([
-  'analytics',
+  'analytics-writer',
   'indexer',
   'piece-retriever',
+  'bad-bits',
   'terminator',
   'workflows',
 ])
